optimize LoginPage, signInForm and signUpForm
