+++ src/app/login/constants.js
@@ -1,0 +1,2 @@
+export const DEFAULT_EMAIL = 'visitor@example.com';
+export const DEFAULT_PASSWORD = 'example';
