+++ src/components/SignInForm/styles.module.scss
@@ -1,0 +1,84 @@
+.pageContainer {
+height: 100%;
+background-color: #f6f7f2;
+}
+
+.loginContainer {
+display: flex;
+flex-direction: column;
+align-items: center;
+justify-content: center;
+height: 100vh;
+background-color: #f6f7f2;
+}
+
+.loginForm {
+display: flex;
+flex-direction: column;
+background-color: white;
+padding: 20px;
+border-radius: 10px;
+box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+min-width: 400px;
+}
+
+.inputContainer {
+margin-bottom: 15px;
+display: flex;
+flex-direction: column;
+}
+
+.label {
+margin-bottom: 5px;
+font-weight: bold;
+color: #333;
+}
+
+.input {
+padding: 10px;
+border: 1px solid #ccc;
+border-radius: 5px;
+}
+
+.button {
+padding: 10px 15px;
+margin-top: 10px;
+background-color: #a7b798;
+color: white;
+border: none;
+border-radius: 5px;
+cursor: pointer;
+text-align: center;
+}
+
+.button:hover {
+background-color: #8f9d7b;
+}
+
+.registerButton {
+background-color: #007bff;
+margin-top: 10px;
+}
+
+.registerButton:hover {
+background-color: #0056b3;
+}
+
+.error {
+color: red;
+font-size: 12px;
+margin-top: 5px;
+}
+.info {
+margin-top: 20px;
+padding: 10px;
+background-color: #f1f1f1;
+border-radius: 5px;
+text-align: center;
+font-size: 14px;
+color: #555;
+}
+
+.info p {
+margin: 5px 0;
+}
